QUIC: ngx_quic_handle_ack_frame_range optimization.
Since all sent packets are sorted in a linked list, we can stop after
the necessary packets are processed in order to avoid processing the
entire linked list. This is necessary to make it possible to increase
the congestion window.

before (max in_flight 1mb):
3.78%  nginx               [.] ngx_quic_handle_ack_frame_range

after (max in_flight 1mb):
0.27%  nginx               [.] ngx_quic_handle_ack_frame_range