QUIC: fixed the sent counter in flow control for STOP_SENDING and RESET_STREAM.
When the client sends STOP_SENDING or RESET_STREAM, packets in flight
and packets in queue will be taken into account in our flow control but
won’t be taken into account in flow control on the client. After many
canceled requests, there might be a large difference and we will not be
able to send more data.

DATA_BLOCKED/STREAM_DATA_BLOCKED is not supported by all clients.